# docs/openapi.yaml
openapi: 3.0.3
info:
  title: dPagès API v1
  version: '1.0'
  description: >
    API interna del sistema de gestión de comandas de dPagès. Fuente única de
    verdad del contrato entre backend y frontend — ver también
    docs/contrato-api.md para las explicaciones de negocio detrás de cada
    regla (por qué el empaquetado exige valores obligatorios, qué es
    kgEditable, etc.).

    Refleja el estado confirmado hasta la capa 11 (detalle de incidencias,
    CORS, alta manual de cliente, dirección de entrega del pedido, fecha y
    observación por línea, código de tarifa) MÁS las decisiones funcionales
    confirmadas con el cliente el 18/08/2026: campos de agrupación de
    artículo (formato, envasado, agrupación de producción), agrupación de
    rendimiento por categoría, Rendiments Porcs, Panell Producció (shape
    cerrado; las fórmulas numéricas todavía no — ver
    docs/contrato-api.md sección 4.10), Panell Obrador por línea individual
    (ya no agrupado por producto), orígenes de pedido como tabla mantenible
    (ya no un enum fijo), y usuarios/roles. Esta versión es SOLO contrato —
    los endpoints nuevos (rendiments-porcs, panells/produccio,
    origens-comanda, usuaris, rols, jo) todavía no están implementados en
    el backend real.
  contact:
    name: VisioFlow Tech
servers:
  - url: http://localhost:8080
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/categories:
    get:
      tags: [Categories]
      summary: Listar categorías de artículo
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Mida'
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dades:
                    type: array
                    items: { $ref: '#/components/schemas/Categoria' }
                  paginacio: { $ref: '#/components/schemas/Paginacio' }
    post:
      tags: [Categories]
      summary: Crear una categoría
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoriaEntrada' }
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Categoria' }
        '400': { $ref: '#/components/responses/Validacio' }
  /api/v1/categories/{id}:
    patch:
      tags: [Categories]
      summary: Actualizar una categoría (parcial)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoriaEntrada' }
      responses:
        '200':
          description: Categoría actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Categoria' }
        '400': { $ref: '#/components/responses/Validacio' }
        '404': { $ref: '#/components/responses/NoTrobat' }
    delete:
      tags: [Categories]
      summary: >-
        Eliminar una categoría. Borrado protegido: si algún artículo activo
        todavía la usa, responde 409 con el recuento en vez de dejar
        artículos con una referencia rota.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204': { description: Eliminada }
        '404': { $ref: '#/components/responses/NoTrobat' }
        '409': { $ref: '#/components/responses/Conflicte' }
...
  /api/v1/rendiments-porcs:
    get:
      tags: [RendimentsPorcs]
      summary: >-
        Listar fichas de rendimiento porcino. Nuevo, confirmado 18/08/2026.
        Sólo devuelve fichas cuyo producto pertenece a una categoría con
        agrupacioRendiment definido (ver descripción del schema RendimentPorc).
      parameters:
        - name: agrupacioRendiment
          in: query
          schema: { type: string, enum: [KG, MAGRE, PAQ] }
        - name: categoria
          in: query
          schema: { type: string }
        - name: agrupacioProduccio
          in: query
          schema: { type: string }
        - name: producte
          in: query
          description: >-
            Filtra por descripció del producte, coincidencia exacta
            (case-insensitive) — regla 3.1 transversal del proyecto, no
            substring.
          schema: { type: string }
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Mida'
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dades:
                    type: array
                    items: { $ref: '#/components/schemas/RendimentPorc' }
                  paginacio: { $ref: '#/components/schemas/Paginacio' }
    post:
      tags: [RendimentsPorcs]
      summary: Crear una ficha de rendimiento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RendimentPorcEntrada' }
      responses:
        '201':
          description: Ficha creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RendimentPorc' }
        '400': { $ref: '#/components/responses/Validacio' }
  /api/v1/rendiments-porcs/{id}:
    patch:
      tags: [RendimentsPorcs]
      summary: >-
        Actualizar una ficha de rendimiento (parcial, edición en línea).
        producteId no se edita una vez creada la fila.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                unitatsPerPorc: { type: string, example: '2.00' }
                kgPerUnitat: { type: string, example: '3.500' }
      responses:
        '200':
          description: Ficha actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RendimentPorc' }
        '400': { $ref: '#/components/responses/Validacio' }
        '404': { $ref: '#/components/responses/NoTrobat' }
    delete:
      tags: [RendimentsPorcs]
      summary: Eliminar una ficha de rendimiento
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204': { description: Eliminada }
        '404': { $ref: '#/components/responses/NoTrobat' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: >
        Requerido en todas las rutas excepto /salut. En desarrollo local con
        AUTH_DISABLED=true no hace falta enviarlo.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
    Pagina:
      name: pagina
      in: query
      schema: { type: integer, default: 1 }
    Mida:
      name: mida
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
  responses:
    NoTrobat:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Validacio:
      description: Error de validación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflicte:
      description: Conflicto (p. ej. pedido congelado)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            codi:
              type: string
              enum: [VALIDACIO, NO_AUTENTICAT, SENSE_PERMIS, NO_TROBAT, CONFLICTE, ERROR_INTERN]
            missatge: { type: string }
            detalls:
              type: array
              items:
                type: object
                properties:
                  camp: { type: string }
                  missatge: { type: string }
    Paginacio:
      type: object
      properties:
        pagina: { type: integer }
        mida: { type: integer }
        total: { type: integer }
        totalPagines: { type: integer }
    Categoria:
      type: object
      properties:
        id: { type: integer }
        nom: { type: string }
        elaboratPorc: { type: boolean }
        agrupacioRendiment:
          type: string
          enum: [KG, MAGRE, PAQ]
          nullable: true
          description: >-
            Null sólo cuando elaboratPorc es false — regla de negocio, no
            falta de dato. Confirmado con el cliente el 18/08/2026.
    CategoriaEntrada:
      type: object
      description: >-
        Cuerpo de POST /categories (nom obligatorio) y PATCH /categories/{id}
        (cuerpo parcial). agrupacioRendiment sólo se acepta si elaboratPorc
        (el enviado, o si no viene, el actual) es true.
      properties:
        nom: { type: string }
        elaboratPorc: { type: boolean, default: false }
        agrupacioRendiment:
          type: string
          enum: [KG, MAGRE, PAQ]
          nullable: true
...
    RendimentPorc:
      type: object
      description: >-
        Ficha de rendimiento por producto — base del cálculo de Panell
        Producció. Nueva, confirmada el 18/08/2026. Sólo aparece si el
        producto pertenece a una categoría con agrupacioRendiment definido
        (ver GET /rendiments-porcs).
      properties:
        id: { type: integer }
        producte:
          type: object
          properties:
            id: { type: integer }
            codi: { type: string, nullable: true }
            descripcio: { type: string }
        agrupacioRendiment:
          type: string
          description: Derivado de producte.categoria.agrupacioRendiment — sólo lectura.
        categoria:
          type: string
          description: Derivado de producte.categoria.nom — sólo lectura.
        agrupacioProduccio:
          type: string
          nullable: true
          description: Derivado de producte.agrupacioProduccio — sólo lectura.
        unitatsPerPorc: { type: string, example: '2.00' }
        kgPerUnitat: { type: string, example: '3.500' }
        pesTotal:
          type: string
          example: '7.000'
          description: Calculado = unitatsPerPorc × kgPerUnitat.
    RendimentPorcEntrada:
      type: object
      required: [producteId, unitatsPerPorc, kgPerUnitat]
      description: >-
        unitatsPerPorc admite hasta 2 decimales, kgPerUnitat hasta 3 (mismas
        escalas que la columna en base).
      properties:
        producteId: { type: integer }
        unitatsPerPorc: { type: string, example: '2.00' }
        kgPerUnitat: { type: string, example: '3.500' }
...